FIX: renaming the train_data to properly signal the architecture
Else, when we add the other architectures, it will be very confusing to have 3 or 4 archives all named train_data. 

This archive is the data gathered during training for the lsx_v2. Other architectures when added should have their own training 
data added with this style as well (train_data_NAMEOFTHEMODEL.json).